+++ pmd-java/src/main/java/net/sourceforge/pmd/lang/java/rule/errorprone/CloseResourceRule.java
@@ -12,9 +12,8 @@ import java.util.HashMap;
 import java.util.HashSet;
 import java.util.List;
 import java.util.Map;
+import java.util.Objects;
 import java.util.Set;
-
-import org.jaxen.JaxenException;
 
 import net.sourceforge.pmd.RuleContext;
 import net.sourceforge.pmd.lang.ast.Node;
@@ -25,10 +24,13 @@ import net.sourceforge.pmd.lang.java.ast.ASTBlock;
 import net.sourceforge.pmd.lang.java.ast.ASTBlockStatement;
 import net.sourceforge.pmd.lang.java.ast.ASTClassOrInterfaceType;
 import net.sourceforge.pmd.lang.java.ast.ASTConstructorDeclaration;
+import net.sourceforge.pmd.lang.java.ast.ASTEqualityExpression;
 import net.sourceforge.pmd.lang.java.ast.ASTExpression;
 import net.sourceforge.pmd.lang.java.ast.ASTFinallyStatement;
 import net.sourceforge.pmd.lang.java.ast.ASTFormalParameters;
 import net.sourceforge.pmd.lang.java.ast.ASTIfStatement;
+import net.sourceforge.pmd.lang.java.ast.ASTImportDeclaration;
+import net.sourceforge.pmd.lang.java.ast.ASTLiteral;
 import net.sourceforge.pmd.lang.java.ast.ASTLocalVariableDeclaration;
 import net.sourceforge.pmd.lang.java.ast.ASTMethodDeclaration;
 import net.sourceforge.pmd.lang.java.ast.ASTMethodOrConstructorDeclaration;
@@ -107,6 +109,7 @@ public class CloseResourceRule extends AbstractJavaRule {
     // keeps track of already reported violations to avoid duplicated violations for the same variable
     private final Set<String> reportedVarNames = new HashSet<>();
 
+    private boolean hasStaticImportObjectsNonNull;
 
     public CloseResourceRule() {
         definePropertyDescriptor(CLOSE_TARGETS_DESCRIPTOR);
@@ -118,6 +121,8 @@ public class CloseResourceRule extends AbstractJavaRule {
 
     @Override
     public void start(RuleContext ctx) {
+        hasStaticImportObjectsNonNull = false;
+
         closeTargets.clear();
         simpleTypes.clear();
         types.clear();
@@ -526,17 +531,87 @@ public class CloseResourceRule extends AbstractJavaRule {
     private boolean isNotConditional(ASTBlock enclosingBlock, Node node, String varName) {
         ASTIfStatement ifStatement = findIfStatement(enclosingBlock, node);
         if (ifStatement != null) {
-            try {
-                // find expressions like: varName != null or null != varName
-                List<?> nodes = ifStatement.findChildNodesWithXPath("Expression/EqualityExpression[@Image='!=']"
-                        + "  [PrimaryExpression/PrimaryPrefix/Name[@Image='" + varName + "']]"
-                        + "  [PrimaryExpression/PrimaryPrefix/Literal/NullLiteral]");
-                return !nodes.isEmpty();
-            } catch (JaxenException e) {
-                throw new RuntimeException(e);
+
+            // find expressions like: varName != null or null != varName
+            // Expression/EqualityExpression[@Image='!=']
+            //   [PrimaryExpression/PrimaryPrefix/Name[@Image='" + varName + "']]
+            //   [PrimaryExpression/PrimaryPrefix/Literal/NullLiteral]
+            ASTEqualityExpression equalityExpr = ifStatement.getCondition().getFirstChildOfType(ASTEqualityExpression.class);
+            if (equalityExpr != null && "!=".equals(equalityExpr.getOperator())) {
+                JavaNode left = equalityExpr.getChild(0);
+                JavaNode right = equalityExpr.getChild(1);
+                
+                if (isVariableAccess(left, varName) && isNullLiteral(right)
+                        || isVariableAccess(right, varName) && isNullLiteral(left)) {
+                    return true;
+                }
             }
+
+            // find method call Objects.nonNull(varName)
+            if (isMethodCall(ifStatement.getCondition())) {
+                ASTPrimaryExpression methodCall = ifStatement.getCondition().getFirstChildOfType(ASTPrimaryExpression.class);
+                ASTPrimaryPrefix prefix = methodCall.getFirstChildOfType(ASTPrimaryPrefix.class);
+                ASTName methodName = prefix.getFirstChildOfType(ASTName.class);
+                if (isObjectsNonNull(methodName)) {
+                    ASTArgumentList arguments = methodCall.getFirstChildOfType(ASTPrimarySuffix.class)
+                            .getFirstDescendantOfType(ASTArgumentList.class);
+                    if (arguments.size() == 1) {
+                        JavaNode firstArgument = arguments.getChild(0);
+                        if (firstArgument.getNumChildren() > 0) {
+                            return isVariableAccess(firstArgument.getChild(0), varName);
+                        }
+                    }
+                }
+            }
+    
+            return false;
         }
         return true;
+    }
+
+    @Override
+    public Object visit(ASTImportDeclaration node, Object data) {
+        if (node.isStatic()) {
+            if ("java.util.Objects".equals(node.getImportedName()) && node.isImportOnDemand()
+                    || "java.util.Objects.nonNull".equals(node.getImportedName()) && !node.isImportOnDemand()) {
+                hasStaticImportObjectsNonNull = true;
+            }
+        }
+        return super.visit(node, data);
+    }
+
+    private boolean isObjectsNonNull(ASTName methodName) {
+        if (methodName == null) {
+            return false;
+        }
+        if (methodName.hasImageEqualTo("Objects.nonNull")) {
+            return methodName.getType() == Objects.class;
+        }
+        if (methodName.hasImageEqualTo("nonNull")) {
+            return hasStaticImportObjectsNonNull;
+        }
+
+        return false;
+    }
+
+    private boolean isVariableAccess(JavaNode node, String varName) {
+        if (node == null || node.getNumChildren() < 1 || node.getChild(0).getNumChildren() < 1) {
+            return false;
+        }
+
+        return node instanceof ASTPrimaryExpression && node.getChild(0) instanceof ASTPrimaryPrefix
+                && node.getChild(0).getChild(0) instanceof ASTName
+                && node.getChild(0).getChild(0).hasImageEqualTo(varName);
+    }
+
+    private boolean isNullLiteral(JavaNode node) {
+        if (node == null || node.getNumChildren() < 1 || node.getChild(0).getNumChildren() < 1) {
+            return false;
+        }
+
+        return node instanceof ASTPrimaryExpression && node.getChild(0) instanceof ASTPrimaryPrefix
+                && node.getChild(0).getChild(0) instanceof ASTLiteral
+                && node.getChild(0).getChild(0).getFirstChildOfType(ASTNullLiteral.class) != null;
     }
 
     private ASTIfStatement findIfStatement(ASTBlock enclosingBlock, Node node) {